+++ puppetmaster/__init__.py
@@ -2,5 +2,5 @@
 
 __all__ = ["__version__"]
 
-__version__ = "0.9.76"
+__version__ = "0.9.77"